fix(info): 压缩信息页工具区与分页布局
Constraint: #170 reopened feedback requires desktop actions in title row, unframed filters, and mobile compact layout

Rejected: keeping filters inside a surface | still consumed vertical space and diverged from expected layout

Confidence: high

Scope-risk: moderate

Directive: Keep desktop info actions in the page title command row and keep filter controls unframed unless the issue is reopened with new visual requirements

Tested: flutter analyze --no-fatal-infos; flutter test test/info_page_layout_test.dart --reporter compact; flutter test test/widget_test.dart --name "移动端安全区" --reporter compact; flutter test --reporter compact

Not-tested: manual Android/iOS device visual inspection
