+++ docs/DESIGN.md
@@ -195,9 +195,9 @@ _initApp()
 - 本地缓存持久化与已读状态管理
 - 设置页仅展示已接入抓取链路的渠道；微信服务号历史枚举仅用于旧缓存兼容，不作为可开启入口
 
-**UI 结构**：操作、刷新进度、搜索和筛选集中在同一个 Fluent surface 中，消息列表和空状态使用独立 surface，减少分散按钮和裸列表带来的层级割裂。
+**UI 结构**：桌面 / 平板端将全部标为已读、刷新官网消息和刷新微信推文放入“信息中心”标题行命令区，内容区首行仅保留搜索、筛选与未读开关的紧凑工具栏，工具栏不使用独立卡片框，消息列表和空状态使用独立 surface，减少非内容区域占高。
 
-**移动端适配**：Compact 宽度下使用紧凑标题 / 搜索 / 图标操作区，次级筛选收进 Fluent 对话框，页面主体保留当前筛选摘要；分页固定为上一页 / 当前页状态 / 下一页单行结构，优先保证消息列表首屏可视高度。
+**移动端适配**：Compact 宽度下使用搜索优先的单行图标操作区，次级筛选收进 Fluent 对话框，页面主体保留当前筛选摘要；分页固定为上一页 / 当前页状态 / 下一页单行结构，优先保证消息列表首屏可视高度。
 
 ### 4.4 快速跳转（QuickLinksPage）
